+++ src/MudBlazor/Components/Chart/Models/Data/ChartSeries.cs
