+++ src/MudBlazor/Components/Chart/Models/Data/LineDisplayType.cs
@@ -1,0 +1,7 @@
+﻿namespace MudBlazor;
+
+public enum LineDisplayType
+{
+    Line,
+    Area,
+}
