--- src/MudBlazor/Components/Chart/Models/LineDisplayType.cs
@@ -1,15 +1,0 @@
-﻿#nullable enable
-namespace MudBlazor;
-
-[Obsolete("Use LineDisplayType instead. This will be removed in a future major version.", false)]
-public enum TimeSeriesDisplayType
-{
-    Line,
-    Area,
-}
-
-public enum LineDisplayType
-{
-    Line,
-    Area,
-}
