+++ src/MudBlazor/Components/Chart/Models/Options/InterpolationOption.cs
