+++ src/MudBlazor/Components/Chart/MudChart.razor
@@ -25,7 +25,7 @@
             }
             @if (ChartType == ChartType.StackedBar)
             {
-                <StackedBar ChartSeries="@ChartSeries" AxisChartOptions="AxisChartOptions" @bind-SelectedIndex="@SelectedIndex" XAxisLabels="@XAxisLabels"></StackedBar>
+                <StackedBar ChartSeries="@ChartSeries" Options="ChartOptions" @bind-SelectedIndex="@SelectedIndex" XAxisLabels="@XAxisLabels"></StackedBar>
             }
             @if (ChartType == ChartType.HeatMap)
             {
